Revamp the representation of operations.
Before this change, operations were used as trait objects throughout
this crate. This had the advantage that it was easy to create define
custom distance measure, since a distance measure was effectively just
a Vec of boxed edit operations.

However, this implementation was subpar from the perspective of crate
users. Since the edit script contained boxed operations:

Vec<Box<EditOperation>>

The trait objects have to be downcases to get the actual edit
operations. This get tedious very quickly.

In this change we switch to static polymorphism, so that we can return
vectors such as:

Vec<LevenshteinOp>

where `LevenshteinOp` is an enum. This makes the crate more usable, at
a higher cost for implementing new measures, since each measure has its
own edit operations.